% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{augmentation}
\alias{augmentation}
\title{Function for data augmentation, go inside a list and append the new pictures}
\usage{
augmentation(data, shift = c(), gamma = c())
}
\arguments{
\item{data}{list of pictures return by f.read.image with options ls.format = T}

\item{shift}{vector of shift values. c(0) means no shifting.}

\item{gamma}{vector of gamma values, c(1) means no attenuation. normalization will be applied after in order to stay in the 0-1 range (note that normalization should be applied anyway for efficienty)}
}
\description{
Function for data augmentation, go inside a list and append the new pictures
}
\examples{
data <- f.read.image(paste0('www/propolis-silicate-',seq(6),'.jpg'),height = 126,Normalize = T,ls.format = T)
augmented <- augmentation(data,shift=c(-5,5),gamma=c(0.5,2))
abind(augmented,along=2) \%>\% raster()
}
\author{
Dimitri Fichou
}

