% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm.R
\name{binary.state}
\alias{binary.state}
\title{Calculate probability real values}
\usage{
binary.state(h)
}
\arguments{
\item{h}{hidden units states}
}
\description{
Calculate probability real values
}
\author{
Xiao Rong
}

