% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.pict.R
\name{chrom.pict}
\alias{chrom.pict}
\title{Function to plot the 3 channel of an array as chromatograms,}
\usage{
chrom.pict(data, x, normalization = T, edge = 0, ...)
}
\arguments{
\item{data}{a 3D array of dim[3] = 3}

\item{x}{the x coordinate to look at}

\item{normalization}{either or not normalisation should happen before plotting the chromatograms}

\item{edge}{to apply a mean and avoid noise}

\item{...}{extra parameters to be passed to the plot function}
}
\description{
Function to plot the 3 channel of an array as chromatograms,
}
\author{
Dimitri Fichou
}
