% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.raster.R
\name{chrom.raster}
\alias{chrom.raster}
\title{Function to plot a picture AND the 3 channel of an array as chromatograms,}
\usage{
chrom.raster(data, x, normalization = F, edge = 0, ...)
}
\arguments{
\item{data}{a 3D array of dim[3] = 3}

\item{x}{the x coordinate to look at}

\item{normalization}{either or not normalisation should happen before plotting the chromatograms}

\item{edge}{allow to apply the mean function over a range to increase teh signal/noise ratio}

\item{...}{extra parameters to be passed to the plot function}
}
\description{
Function to plot a picture AND the 3 channel of an array as chromatograms,
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
data \%>\% chrom.raster(x=140)
}
\author{
Dimitri Fichou
}
