% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterize.R
\name{clusterize}
\alias{clusterize}
\title{function to clusterize the code of the model}
\usage{
clusterize(data, margin, transform, dimension)
}
\arguments{
\item{data}{a data frame or a matrix}

\item{margin}{the margin used for deconstruction}

\item{transform}{was transform used for deconstruction}

\item{dimension}{the dimension of the original array}
}
\description{
function to clusterize the code of the model
}
\author{
Dimitri Fichou
}

