% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.new.color.R
\name{create.new.color}
\alias{create.new.color}
\title{Function to calculate the euclidean distance of each pixel to one selected pixel}
\usage{
create.new.color(data, x, y, func)
}
\arguments{
\item{data}{a 3D array, basically create by the function f.read.image}

\item{x}{the x coordonate of the picture (x is the second dimension)}

\item{y}{the y coordonate of the picture (y is the first dimension and start from the top, dont ask why)}
}
\description{
Function to calculate the euclidean distance of each pixel to one selected pixel
}
\examples{
data <- f.read.image('www/rTLC_demopicture.JPG',format='jpeg',native=F) \%>\% redim.array(256)
par(mfrow=c(2,1),oma=c(0,0,0,0),mar=c(0,0,1,0),xaxt='n',yaxt='n')
x=180;y=70;
data \%>\% raster(main='original')
text(x=x,y=y,labels='X',col='red')
data \%>\% create.new.color(x,y) \%>\% normalize \%>\% raster(main='New color')
text(x=x,y=y,labels='X',col='red')
}
\author{
Dimitri Fichou
}
