% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_train.R
\name{dbn.down}
\alias{dbn.down}
\title{Predict visible unit from top hidden layer}
\usage{
dbn.down(dbn, h, round = 10)
}
\arguments{
\item{dbn}{model}

\item{h}{state of hidden unit}

\item{round}{number of time to cross the network at each layer}
}
\description{
Predict visible unit from top hidden layer
}
\author{
Xiao Rong
}
