% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_train.R
\name{dbn.train}
\alias{dbn.train}
\title{Training a Deep Belief Net}
\usage{
dbn.train(x, hidden = c(10, 10), numepochs = 3, batchsize = 100,
  learningrate = 0.8, learningrate_scale = 1, momentum = 0.5,
  visible_type = "bin", hidden_type = "bin", cd = 1)
}
\arguments{
\item{x}{matrix of x values for examples}

\item{hidden}{number of hidden units}

\item{numepochs}{number of iteration for samples  Default is 3.}

\item{batchsize}{size of mini-batch. Default is 100.}

\item{learningrate}{learning rate for gradient descent. Default is 0.8.}

\item{learningrate_scale}{learning rate will be mutiplied by this scale after every iteration. Default is 1 .}

\item{momentum}{momentum for gradient descent. Default is 0.5 .}

\item{visible_type}{activation function of input unit.Only support "sigm" now}

\item{hidden_type}{activation function of hidden unit.Only support "sigm" now}

\item{cd}{number of iteration for Gibbs sample of CD algorithm.}
}
\description{
Training a Deep Belief Net
}
\author{
Xiao Rong
}

