% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_train.R
\name{dbn.up}
\alias{dbn.up}
\title{Predict top layer hidden unit from visible units}
\usage{
dbn.up(dbn, v, round = 10)
}
\arguments{
\item{dbn}{model}

\item{v}{matrix of x values for examples}

\item{round}{number of time to cross the network at each layer}
}
\description{
Predict top layer hidden unit from visible units
}
\author{
Xiao Rong
}

