% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstruct.convol.R
\name{deconstruct.convol}
\alias{deconstruct.convol}
\title{function to deconstruct an array,}
\usage{
deconstruct.convol(data, margin, transform, conv_width = 0)
}
\arguments{
\item{data}{a data frame or a matrix}

\item{margin}{the margin used for deconstruction}

\item{transform}{was transform used for deconstruction}

\item{conv_width}{number of pixels to take around, 0 mean take only 1}
}
\description{
function to deconstruct an array,
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
print(dim(data))
decon <- data \%>\% deconstruct.convol(margin =3 ,transform = F, conv_width =2 )
print(dim(decon))
}
\author{
Dimitri Fichou
}

