% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm.R
\name{do.rbm.train}
\alias{do.rbm.train}
\title{do the rbm training, mean to be use inside rbm.train}
\usage{
do.rbm.train(rbm, batch_x, s)
}
\arguments{
\item{rbm}{an rbm object}

\item{batch_x}{batch_x}

\item{s}{s}
}
\description{
do the rbm training, mean to be use inside rbm.train
}
\author{
Xiao Rong
}
