% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{evolve}
\alias{evolve}
\title{function to reconstruct an array from a df by evolving one of the dimension}
\usage{
evolve(data, margin, transform, dimension)
}
\arguments{
\item{data}{a data frame or a matrix}

\item{margin}{the margin used for deconstruction}

\item{transform}{was transform used for deconstruction}

\item{dimension}{the dimension of the original array}
}
\description{
function to reconstruct an array from a df by evolving one of the dimension
}
\author{
Dimitri Fichou
}
