% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.count.successif.R
\name{f.count.successif}
\alias{f.count.successif}
\title{function to count the number of successif occurence in a vector}
\usage{
f.count.successif(vec)
}
\arguments{
\item{vec}{a vector}
}
\description{
function to count the number of successif occurence in a vector
}
\author{
Dimitri Fichou
}

