% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.eat.image.R
\name{f.eat.image}
\alias{f.eat.image}
\title{Function to extract chromatograms from an 3d array, or a 2d matrix}
\usage{
f.eat.image(data, conv = "linomat", largeur = 200, dist.gauche = 20,
  band = 6, ecart = 2, tolerance = 1, cropping = 0,
  nbr.band = NULL, plotting = F)
}
\arguments{
\item{data}{input array or matrix}

\item{conv}{convention for the dimensions, should the dimension correspond to the left part of the band or the center of the band}

\item{largeur}{plate width in mm}

\item{dist.gauche}{distance from the left side of the plate in mm}

\item{band}{band widthin mm}

\item{ecart}{gap between band in mm}

\item{tolerance}{mm to remove on each side of the band in order to take only the center}

\item{cropping}{mm to remove on the left part of the plate in the largeur and dist.gauche arameter in order to counter the intempestive camag cropping}

\item{nbr.band}{if not specified, the software assume it itself, otherwise will force the number of band, could lead to an error if not enough space}

\item{plotting}{will plot the raster to check the extraction}
}
\description{
Function to extract chromatograms from an 3d array, or a 2d matrix
}
\examples{
data <- f.read.image('www/rTLC_demopicture.JPG',height=256)
data <- f.eat.image(data,conv="linomat",largeur = 200,dist.gauche=20,band=6,ecart=2,tolerance=1,cropping = 0,nbr.band=20,plotting=T)
data \%>\% raster()
}
\author{
Dimitri Fichou
}
