% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.integrate.R
\name{f.integrate}
\alias{f.integrate}
\title{function to extract chromato, corret baseline, integrate peak}
\usage{
f.integrate(data, dim.table, negatif = F, correct.baseline = F)
}
\arguments{
\item{data}{an array}

\item{dim.table}{a matrix with a row per peak to extract/integrate, as column respectively xmin, xmax, ymin, ymax}

\item{negatif}{apply a transformation to put the data in negatif, needed for white light pictures and for some extracted featuress from the network}

\item{correct.baseline}{use or not the rolling ball algorithm to correct the baseline}

\item{...}{Arguments to be passed to methods}
}
\description{
function to extract chromato, corret baseline, integrate peak
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
dim.table <- rbind(c(1,2,50,60,100,110),c(1,2,70,80,100,110))
f.integrate(data,dim.table,F,T)
}
\author{
Dimitri Fichou
}

