% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.integrate_extracted.R
\name{f.integrate_extracted}
\alias{f.integrate_extracted}
\title{function to integrate peaks from 3d chromato array}
\usage{
f.integrate_extracted(data, channel, negatif = F, correct.baseline = F,
  allign = F, start, end, wm = 50, ws = 5, plotting = F)
}
\arguments{
\item{data}{an array as returned by the function f.eat.image}

\item{channel}{numeric for the channel to use (1 for red, 2 for green and 3 for blue)}

\item{negatif}{apply a transformation to put the data in negatif, needed for white light pictures and for some extracted featuress from the network}

\item{correct.baseline}{use or not the rolling ball algorithm to correct the baseline}

\item{start}{starting pixel index}

\item{end}{stopping pixel index}

\item{wm}{Width of local window for minimization/maximization}

\item{ws}{Width of local window for smoothing}

\item{plotting}{should a plot be display to check the integration}

\item{align}{boolean, use or not ptw to align the peak to the first sample}

\item{...}{Arguments to be passed to methods}
}
\description{
function to integrate peaks from 3d chromato array
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
## incoming
}
\author{
Dimitri Fichou
}
