% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.read.image.multiple.R
\name{f.read.image.multiple}
\alias{f.read.image.multiple}
\title{Deprecated. Function to read multiple image file}
\usage{
f.read.image.multiple(source, native = F, height, Normalize = F,
  ls.format = F)
}
\arguments{
\item{source}{the path of the file}

\item{native}{should the data be return natively or as a 3D array}

\item{height}{redimension the file}

\item{ls.format}{use to return a list instead of an array}
}
\description{
Deprecated. Function to read multiple image file
}
\examples{
data <- f.read.image('www/rTLC_demopicture.JPG',format='jpeg',native=F,height=256)
data \%>\% raster()
}
\author{
Dimitri Fichou
}

