% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.3D.R
\name{layer.3D}
\alias{layer.3D}
\title{function to plot a matrix}
\usage{
layer.3D(data)
}
\arguments{
\item{data}{a matrix}
}
\description{
function to plot a matrix
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
layer.3D(data[,,1])
}
\author{
Dimitri Fichou
}

