% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.distance.R
\name{map.kohonen}
\alias{map.kohonen}
\title{Modification of the map.kohonen function of the kohonen package in order to return the distance}
\usage{
\method{map}{kohonen}(x, newdata, whatmap = NULL, weights,
  scale.distances = (nmaps > 1), ...)
}
\arguments{
\item{x}{x param}

\item{newdata}{newdata param}

\item{whatmap}{whatmap param}

\item{weights}{weights param}

\item{scale.distances}{scale distance param}
}
\description{
Modification of the map.kohonen function of the kohonen package in order to return the distance
}
\author{
Dimitri Fichou
}

