% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.distance.R
\name{map}
\alias{map}
\title{Modification of the map function of the kohonen package in order to return the distance}
\usage{
map(x, ...)
}
\arguments{
\item{x}{x param}
}
\description{
Modification of the map function of the kohonen package in order to return the distance
}
\author{
Dimitri Fichou
}

