% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Function to normalize array or matrix}
\usage{
normalize(data, hidden = F, ref = NULL)
}
\arguments{
\item{data}{a 3D array or a matrix}

\item{hidden}{option to normalize in other dimension than the dimension 3}

\item{ref}{optionnal picture to keep the max and min values}
}
\description{
Function to normalize array or matrix
}
\examples{
data <- f.read.image('www/rTLC_demopicture.JPG',format='jpeg',native=F) \%>\% redim.array(256)
data \%>\% normalize \%>\% raster
}
\author{
Dimitri Fichou
}
