% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_figure.R
\name{raster_figure}
\alias{raster_figure}
\title{function to plot several figures, for publication}
\usage{
raster_figure(data, lab = NULL, ...)
}
\arguments{
\item{data}{list of image that will be redimension in height and bind in the second dimension}
}
\description{
function to plot several figures, for publication
}
\author{
Dimitri Fichou
}
