% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.convol.R
\name{reconstruct.convol}
\alias{reconstruct.convol}
\title{function to reconstruct an array from a matrix,}
\usage{
reconstruct.convol(data, margin, transform, dimension, conv_width = 0,
  take_center = F, return.ls = F)
}
\arguments{
\item{data}{a matrix}

\item{margin}{the margin used for deconstruction}

\item{transform}{was transform used for deconstruction}

\item{dimension}{dimension of the original array}

\item{conv_width}{number of pixels to take around, 0 mean take only 1}

\item{take_center}{in case of convolution, should the principal pixel only be taken for reconstruction or should each array be shifted and the total mean computed}
}
\description{
function to reconstruct an array from a matrix,
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
decon <- data \%>\% deconstruct.convol(margin =3 ,transform = F, conv_width =2 )
recon <- decon \%>\% reconstruct.convol(margin =3 ,transform = F,dimension=dim(data), conv_width =2)
}
\author{
Dimitri Fichou
}

