% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{function to reconstruct an array from a matrix,}
\usage{
reconstruct(data, margin, transform, dimension)
}
\arguments{
\item{data}{a matrix}

\item{margin}{the margin used for deconstruction}

\item{transform}{was transform used for deconstruction}

\item{dimension}{dimension of the original array}
}
\description{
function to reconstruct an array from a matrix,
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',native=F,format = 'jpeg',height=128)
decon <- data \%>\% deconstruct.convol(margin =3 ,transform = F, conv_width =2 )
recon <- decon \%>\% reconstruct(margin =3 ,transform = F, dim(data))
}
\author{
Dimitri Fichou
}

