% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim.array.R
\name{redim.array}
\alias{redim.array}
\title{Function to redimension a 3D array}
\usage{
redim.array(data, height.out)
}
\arguments{
\item{data}{3D array}

\item{height.out}{number of pixel in height (first dimension of the 3D array)}
}
\description{
Function to redimension a 3D array
}
\examples{
data <- f.read.image('www/rTLC_demopicture.JPG',format='jpeg',native=F)
data <- data \%>\% redim.array(256)
data \%>\% raster()
}
\author{
Dimitri Fichou
}
