% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.tlc_RBM_classification.R
\name{run.tlc_RBM_classification}
\alias{run.tlc_RBM_classification}
\title{Function to launch the tlc_RBM_classification app}
\usage{
run.tlc_RBM_classification(port = NULL)
}
\arguments{
\item{port}{if not NULL will deploy on the local network on this port, just look for the IP of the host and go for example on 192.168.1.20:port to access the app from anywhere in the network}
}
\description{
Function to launch the tlc_RBM_classification app
}
\examples{
run.tlc_RBM_classification()
}
\author{
Dimitri Fichou
}

