% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.tlc_var_selector.R
\name{run.tlc_var_selector}
\alias{run.tlc_var_selector}
\title{Function to launch the tlc_var_selector app}
\usage{
run.tlc_var_selector(port = NULL)
}
\arguments{
\item{port}{if not NULL will deploy on the local network on this port, just look for the IP of the host and go for example on 192.168.1.20:port to access the app from anywhere in the network}
}
\description{
Function to launch the tlc_var_selector app
}
\examples{
run.tlc_var_selector()
}
\author{
Dimitri Fichou
}

