% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.tlcbrain.R
\name{run.tlcbrain}
\alias{run.tlcbrain}
\title{Function to launch the tlc-brain app}
\usage{
run.tlcbrain(port = NULL)
}
\arguments{
\item{port}{if not NULL will deploy on the local network on this port, just look for the IP of the host and go for example on 192.168.1.20:port to access the app from anywhere in the network}
}
\description{
Function to launch the tlc-brain app
}
\examples{
run.tlcbrain()
}
\author{
Dimitri Fichou
}

