% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Function to create the data base on mongodb, not working yet}
\usage{
saveData(data)
}
\arguments{
\item{data}{3D array}
}
\description{
Function to create the data base on mongodb, not working yet
}
\author{
Dimitri Fichou
}

