% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm.R
\name{sigm}
\alias{sigm}
\title{function for sigm unit}
\usage{
sigm(x)
}
\arguments{
\item{x}{unit value}
}
\description{
function for sigm unit
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
data \%>\% sigm \%>\% raster
}
\author{
Xiao Rong
}

