% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.cluster.plot.R
\name{SOM.cluster.plot.picture}
\alias{SOM.cluster.plot.picture}
\title{Function to plot a picture}
\usage{
SOM.cluster.plot.picture(data, main = "")
}
\arguments{
\item{data}{a 3D array}

\item{main}{title of the picture}
}
\description{
Function to plot a picture
}
\author{
Dimitri Fichou
}
