% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.cluster.plot.R
\name{SOM.cluster.plot}
\alias{SOM.cluster.plot}
\title{Function to plot an object return by the SOM.cluster function}
\usage{
SOM.cluster.plot(data, xcoord = NULL, ycoord = NULL, index.original = 1,
  index.recon = 1)
}
\arguments{
\item{data}{list return by the SOM.cluster function}

\item{xcoord}{xcoord of the point to look at}

\item{ycoord}{ycoord of te point to look at}

\item{index.original}{index of the original picture in the list}

\item{index.recon}{index of the reconstruct picture in the list}
}
\description{
Function to plot an object return by the SOM.cluster function
}
\author{
Dimitri Fichou
}
