% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.cluster.R
\name{SOM.cluster}
\alias{SOM.cluster}
\title{Function to apply a pipeline of deconstruction, learning, reconstruction on a picture}
\usage{
SOM.cluster(data, margin, transform = c(F), grid.x = c(3),
  grid.y = c(3), topo = c("hexagonal"), toroidal = c(F),
  rlen = c(100), alpha.1 = c(0.05), alpha.2 = c(0.01),
  radius = c("Default"), action = c("reconstruct"),
  count.successif.1 = rep(0, length(margin)),
  count.successif.2 = rep(100, length(margin)), use = length(margin))
}
\arguments{
\item{data}{a 3D array, create by the function f.read.image}

\item{margin}{the margin to apply the deconstruction}

\item{transform}{should the array be transformed after deconstruction (interesting value are: margin=3-transform=F and margin=2-transform=T)}

\item{grid.x}{x width of the kohonen map}

\item{grid.y}{y width of the kohonen map}

\item{topo}{topo of the kohonen map}

\item{toroidal}{should the map be toroidal}

\item{rlen}{number of iteration}

\item{alpha.1}{learning rate begining}

\item{alpha.2}{learning rate end}

\item{radius}{radius of the neighbourhood}

\item{action}{what the process should return for the next step, choices are: reconstruct, evolve, clusterize, original, original_noirci.}

\item{count.successif.1}{for original_noirci what is the minimum of successive sample to be consider as sample and note noise or interband}

\item{count.successif.2}{for original_noirci what is the maximum of successive sample to be consider as sample and note noise or interband}

\item{use}{use to skip a process}
}
\description{
Function to apply a pipeline of deconstruction, learning, reconstruction on a picture
}
\examples{
data <- f.read.image('www/rTLC_demopicture.JPG',format='jpeg',native=F) \%>\% redim.array(256)
model <- SOM.cluster(data,margin=c(3),transform=F,action='evolve')
str(model)
model$data.recon[[1]][,,1] \%>\% raster
}
\author{
Dimitri Fichou
}
