% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_state_raster.R
\name{unit_state_raster}
\alias{unit_state_raster}
\title{function to plot the image of a given hidden unit state for pixel studies}
\usage{
unit_state_raster(data, model, conv, main = F, normalize = F)
}
\arguments{
\item{data}{a vector of length the number of hidden unit in the network}

\item{model}{the model used}

\item{conv}{conv used for deconstruction}

\item{main}{should the hidden state be plotted}

\item{normalize}{should the windows be normalized}
}
\description{
function to plot the image of a given hidden unit state for pixel studies
}
\author{
Dimitri Fichou
}
