% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_state_raster.R
\name{vec2patch}
\alias{vec2patch}
\title{function to plot the image of a given hidden unit state for pixel studies}
\usage{
vec2patch(vec, conv, normalize = F)
}
\arguments{
\item{vec}{a vector of length the number of hidden unit in the network}

\item{conv}{conv used for deconstruction}

\item{normalize}{should the windows be normalized}
}
\description{
function to plot the image of a given hidden unit state for pixel studies
}
\author{
Dimitri Fichou
}
