% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.click.kohonen.R
\name{which.click.kohonen}
\alias{which.click.kohonen}
\title{Function to know which cell was click, used in the shiny apps}
\usage{
which.click.kohonen(grid.x, grid.y, topo, click.x, click.y)
}
\arguments{
\item{grid.x}{horizontal size of the map}

\item{grid.y}{vertical size of the map}

\item{topo}{topo of the map}

\item{click.x}{x coordonate of the click}

\item{click.y}{y coordonate of the click}
}
\description{
Function to know which cell was click, used in the shiny apps
}
\author{
Dimitri Fichou
}
