% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster}
\alias{raster}
\title{function to plot an image}
\usage{
raster(data, main = "", Title.dim = F, ...)
}
\arguments{
\item{data}{an array, or a matrix}

\item{main}{an optionnal title}

\item{Title.dim}{should the dimension be indicated on the plot}

\item{...}{Arguments to be passed to methods}
}
\description{
function to plot an image
}
\examples{
data <- f.read.image(source='www/rTLC_demopicture.JPG',height=128)
data \%>\% raster(main='HPTLC plate',Title.dim=T)
}
\author{
Dimitri Fichou
}

